+++ test/transport/local/local_transport_cryptosign_test.dart
@@ -1,22 +1,31 @@
 import 'package:connectanum/authentication.dart';
 import 'package:connectanum/connectanum.dart';
+import 'package:connectanum/src/authentication/cryptosign/pkcs8.dart';
 import 'package:test/test.dart';
+
+import '../../authentication/cryptosign/keys.dart';
 
 void main() {
   group('Local transport authentication test', () {
     test('success', () async {
+      var authentication = CryptosignAuthentication.fromPuttyPrivateKey(
+          MockKeys.ed25519Ppk.value);
       var client = Client(
           realm: "com.connectanum",
-          authenticationMethods: [CraAuthentication("Richard")],
+          authenticationMethods: [authentication],
           authId: "Burkhardt",
-          transport: LocalTransport(authenticationPassword: "Richard"));
+          transport:
+              LocalTransport(authenticationKey: authentication.privateKey));
       var session = await client.connect().first;
       expect(session, isNotNull);
     });
     test('fail', () async {
       var client = Client(
           realm: "com.connectanum",
-          authenticationMethods: [CraAuthentication("Richard")],
+          authenticationMethods: [
+            CryptosignAuthentication.fromPuttyPrivateKey(
+                MockKeys.ed25519Ppk.value)
+          ],
           authId: "Burkhardt",
           transport: LocalTransport());
       AbstractMessage? errorMessage;